requirements.txt
PyYAML>=6.0
Pillow>=10.0
matplotlib>=3.8
